refactor(security): move CORS middleware to common middlewares directory
- Extract corsMiddleware from config to dedicated middleware file
- Maintain same functionality with proper error handling
- Improve code organization and separation of concerns
